+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Unreleased
 
 ### Fixed
+- Fixed a bug where the vCard was not downloading correctly.
 - Fixed an error that could occur when creating a vCard with multiple URLs. ([#8](https://github.com/nfourtythree/craft3-vcard/pull/8))
 
 ## 1.1.0 - 2020-05-20 [CRITICAL]